Add files via upload
Fix: Timer and tries count issue

Two issues were identified by reviewer:
a) counting of tries by user - duplicity of count while clicking on the same card
b) timer glitches with repeated clicking

A) Counting of tries
The tweak was simple for this - on line 105 of musicMemory.js file. If statement that was inserted "if (cardA.id != cardB.id) {"

B) Timer Issue
The issue was resolved by hiding the button (startTheGame) after being clicked. This however proved more difficulty in practicality than theory because of inherited/defaulted display "in-line block" that was shadowing the button. To address:

1. In CSS file:
a) Separate "buttonDisappearing" class was created at line 253.
b) Default display was set to "none".
c) "!Important" was added to display to allow override.

2. In musicMemory.js file:
a) In the fill function on line 50, the display was set to make the button visible:
document.getElementById("startTheGame").style.setProperty("display", "inline-block");
b) In startGame function on line 247, the display was set to "none" to disappear when the game starts.
document.getElementById("startTheGame").style.setProperty("display", "none")

3. In the musicMemory.html file, the class was made/changed in the disappearing start button to allow the button to disappear:
Play Game